dav_util.c: move dav_get_validators(), dav_store_resource() and precondition enums from http_dav.c to dav_util.c
