+++ imap/http_dav.h
@@ -172,96 +172,6 @@ enum {
                                            schedule-send-reply,
                                            schedule-send-freebusy) */
 
-/* Index into preconditions array */
-enum {
-    /* WebDAV (RFC 4918) preconditions */
-    DAV_PROT_PROP = 1,
-    DAV_BAD_LOCK_TOKEN,
-    DAV_NEED_LOCK_TOKEN,
-    DAV_LOCKED,
-    DAV_FINITE_DEPTH,
-
-    /* WebDAV Versioning (RFC 3253) preconditions */
-    DAV_SUPP_REPORT,
-    DAV_RES_EXISTS,
-
-    /* WebDAV ACL (RFC 3744) preconditions */
-    DAV_NEED_PRIVS,
-    DAV_NO_INVERT,
-    DAV_NO_ABSTRACT,
-    DAV_SUPP_PRIV,
-    DAV_RECOG_PRINC,
-    DAV_ALLOW_PRINC,
-    DAV_GRANT_ONLY,
-
-    /* WebDAV Quota (RFC 4331) preconditions */
-    DAV_OVER_QUOTA,
-    DAV_NO_DISK_SPACE,
-
-    /* WebDAV Extended MKCOL (RFC 5689) preconditions */
-    DAV_VALID_RESTYPE,
-
-    /* WebDAV Sync (RFC 6578) preconditions */
-    DAV_SYNC_TOKEN,
-    DAV_OVER_LIMIT,
-
-    /* CalDAV (RFC 4791) preconditions */
-    CALDAV_SUPP_DATA,
-    CALDAV_VALID_DATA,
-    CALDAV_VALID_OBJECT,
-    CALDAV_SUPP_COMP,
-    CALDAV_LOCATION_OK,
-    CALDAV_UID_CONFLICT,
-    CALDAV_SUPP_FILTER,
-    CALDAV_VALID_FILTER,
-    CALDAV_SUPP_COLLATION,
-    CALDAV_MAX_SIZE,
-
-    /* RSCALE (RFC 7529) preconditions */
-    CALDAV_SUPP_RSCALE,
-
-    /* Time Zones by Reference (RFC 7809) preconditions */
-    CALDAV_VALID_TIMEZONE,
-
-    /* Managed Attachments (draft-ietf-calext-caldav-attachments) preconditions */
-    CALDAV_VALID_MANAGEDID,
-
-    /* Bulk Change (draft-daboo-calendarserver-bulk-change) preconditions */
-    CALDAV_CTAG_OK,
-
-    /* CalDAV Scheduling (RFC 6638) preconditions */
-    CALDAV_VALID_SCHED,
-    CALDAV_VALID_ORGANIZER,
-    CALDAV_UNIQUE_OBJECT,
-    CALDAV_SAME_ORGANIZER,
-    CALDAV_ALLOWED_ORG_CHANGE,
-    CALDAV_ALLOWED_ATT_CHANGE,
-    CALDAV_DEFAULT_NEEDED,
-    CALDAV_VALID_DEFAULT,
-
-    /* iSchedule (draft-desruisseaux-ischedule) preconditions */
-    ISCHED_UNSUPP_VERSION,
-    ISCHED_UNSUPP_DATA,
-    ISCHED_INVALID_DATA,
-    ISCHED_INVALID_SCHED,
-    ISCHED_ORIG_MISSING,
-    ISCHED_MULTIPLE_ORIG,
-    ISCHED_ORIG_INVALID,
-    ISCHED_ORIG_DENIED,
-    ISCHED_RECIP_MISSING,
-    ISCHED_RECIP_MISMATCH,
-    ISCHED_VERIFICATION_FAILED,
-
-    /* CardDAV (RFC 6352) preconditions */
-    CARDDAV_SUPP_DATA,
-    CARDDAV_VALID_DATA,
-    CARDDAV_UID_CONFLICT,
-    CARDDAV_LOCATION_OK,
-    CARDDAV_SUPP_FILTER,
-    CARDDAV_SUPP_COLLATION,
-    CARDDAV_MAX_SIZE,
-};
-
 /* Preference bits */
 enum {
     PREFER_MIN    = (1<<0),
@@ -688,18 +598,11 @@ int dav_parse_req_target(struct transaction_t *txn,
                          struct meth_params *params);
 int calcarddav_parse_path(const char *path, struct request_target_t *tgt,
                           const char *mboxprefix, const char **resultstr);
-int dav_get_validators(struct mailbox *mailbox, void *data,
-                       const char *userid, struct index_record *record,
-                       const char **etag, time_t *lastmod);
 modseq_t dav_get_modseq(struct mailbox *mailbox,
                         void *data, const char *userid);
 int dav_check_precond(struct transaction_t *txn, struct meth_params *params,
                       struct mailbox *mailbox, const void *data,
                       const char *etag, time_t lastmod);
-int dav_store_resource(struct transaction_t *txn,
-                       const char *data, size_t datalen,
-                       struct mailbox *mailbox, struct index_record *oldrecord,
-                       modseq_t createdmodseq, strarray_t *imapflags);
 int dav_premethod(struct transaction_t *txn);
 unsigned get_preferences(struct transaction_t *txn);
 struct mime_type_t *get_accept_type(const char **hdr, struct mime_type_t *types);
